requirements.txt
# Necessary libraries.

mysql-connector-python==8.3.0
PyJWT==2.8.0